Work to get energy flow labels better
The .label was interfering with the css label for bootstrap. I basically
just upgraded mermaid js locally and got rid of the joplin stuff and
then manually overrode the label things that were causing trouble but
only for mermaid divs. Seems to be a lot better Still has lots of space
vertically.