Reduced confusing 'thermal' terminology.
The term 'thermal' wrt thermal reactors having neutrons in thermal
equilibrium with their surroundings is confusing to most laypeople and
overused in the nuclear industry. This edit goes through uses
of the nomenclature across the site and replaces it as appropriate
with less confusing terminology.

See: https://twitter.com/whatisnuclear/status/1464285309152075781