Add a few more section navs to large pages.
These are supposed to help people navigate the big pages.
I removed the toc: true thing because we weren't using
it. Instead we're just using kramdown's native toc feature.

I wanted it to be more responsive and stuff but this is a start
at least.

Gotta learn about grid layouts more.