Clean up and extend old video pages and data.
In preparation for new vid announcements, we're doing some cleanups
here. We moved the zppr page to be a news post similar to the hallam
announcement. We'll plan to do news posts for each new vid we
get digitized. Meanwhile, we upgraded the old videos page to
first highlight all the one's we've been involved in digitizing,
then show those others have digitized, and then show the wish list.

Will set up a server-side redirect to maintain the previous zppr
url.