Add EXIF descriptions to all the BONUS files and load them as ALT
This allows me to set descriptions in JPG files themselves, and then
set the alt tags on images directly from the jpg metadata.
I'm hoping this way it'll be easier to reuse the alt text if
the jpgs are used elsewhere.

Also just adding alt tags is nice for e.g. people who can't
see the images.