Add Museum page to collect all of our digitizations/scans.
Two things happened:

* We started getting scans of pamphlets in addition to
  just having scanned films
* The list of scanned films slapped onto the end of each film
  was getting too long and ugly.

So I decided to make a new page that would summarize all the things
we've digitized with little card previews. It includes the films
and pamphlets and photos and the rickover memo.

This also let me pull a few weird history things out of the main
menu, as they were a bit weird there.

Having done this, there was no point in having the 'digitized by
us' section on the old-videos page either, so I collapsed
the digitized film sections into one. I did add a little badge
for the ones we were involved in b/c fun.

Hopefully this will be a bit more scalable if/when we get
more stuff to digitize. Actually I have 3 pamphlets on hand
that need posting right now!

Along the way I found a few spelling errors and incorrect images.